# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: FitnessLab API
  description: API for FitnessLab management system
  version: 1.0.0
servers: []
...
paths:
...
  /api/contracts/customer/{customerId}:
    get:
      tags:
      - Contracts
      operationId: getContractsByCustomerId
      summary: Get all Contracts for a customer
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of contracts for the customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractDetailDto'
  /api/contracts/{contractId}:
    get:
      tags:
      - Contracts
      operationId: getContractById
      summary: Get contract by ID
      parameters:
      - name: contractId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Contract found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDetailDto'
        '404':
          description: Contract not found
  /api/contracts/{contractId}/pause:
    post:
      tags:
      - Contracts
      operationId: pauseContract
      summary: Pause a contract
      parameters:
      - name: contractId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PauseContractRequest'
      responses:
        '200':
          description: Contract paused successfully
  /api/contracts/{contractId}/resume:
    post:
      tags:
      - Contracts
      operationId: resumeContract
      summary: Resume a paused contract
      parameters:
      - name: contractId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Contract resumed successfully
  /api/invoices:
    get:
      tags:
      - Invoices
      operationId: getInvoices
      summary: Get all invoices, optionally filtered by status
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - PAID
          - OVERDUE
          - CANCELLED
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceDto'
  /api/invoices/customer/{customerId}:
    get:
      tags:
      - Invoices
      operationId: getInvoicesByCustomerId
      summary: Get all invoices for a customer
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of customer invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceDto'
  /api/invoices/{invoiceId}:
    get:
      tags:
      - Invoices
      operationId: getInvoiceById
      summary: Get invoice by ID
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDto'
        '404':
          description: Invoice not found
  /api/invoices/{invoiceId}/pay:
    post:
      tags:
      - Invoices
      operationId: markAsPaid
      summary: Mark an invoice as paid
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice marked as paid
  /api/invoices/{invoiceId}/mark-overdue:
    post:
      tags:
      - Invoices
      operationId: markAsOverdue
      summary: Mark an invoice as overdue
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice marked as overdue
  /api/invoices/{invoiceId}/cancel:
    post:
      tags:
      - Invoices
      operationId: cancelInvoice
      summary: Cancel an invoice
      parameters:
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelInvoiceRequest'
      responses:
        '200':
          description: Invoice cancelled successfully
components:
  schemas:
...
    ContractDetailDto:
      type: object
      required:
      - contractId
      - customerId
      - productVariantId
      - bookingId
      - status
      - sessionsUsed
      - pauseHistory
      - canBePaused
      properties:
        contractId:
          type: string
        customerId:
          type: string
        productVariantId:
          type: string
        productName:
          type: string
          nullable: true
        bookingId:
          type: string
        status:
          type: string
        validity:
          $ref: '#/components/schemas/DateRangeDto'
          nullable: true
        sessionsTotal:
          type: integer
          nullable: true
        sessionsUsed:
          type: integer
        pauseHistory:
          type: array
          items:
            $ref: '#/components/schemas/PauseHistoryEntryDto'
        canBePaused:
          type: boolean
    DateRangeDto:
      type: object
      required:
      - start
      - end
      properties:
        start:
          type: string
          format: date
        end:
          type: string
          format: date
    PauseHistoryEntryDto:
      type: object
      required:
      - pauseRange
      - reason
      properties:
        pauseRange:
          $ref: '#/components/schemas/DateRangeDto'
        reason:
          type: string
    PauseContractRequest:
      type: object
      required:
      - startDate
      - endDate
      - reason
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
          enum:
          - MEDICAL
          - VACATION
          - OTHER
    InvoiceDto:
      type: object
      required:
      - invoiceId
      - customerId
      - customerName
      - bookingId
      - amount
      - dueDate
      - status
      - isInstallment
      properties:
        invoiceId:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        bookingId:
          type: string
        amount:
          type: number
          format: decimal
        dueDate:
          type: string
          format: date
        status:
          type: string
        isInstallment:
          type: boolean
        installmentNumber:
          type: integer
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
    CancelInvoiceRequest:
      type: object
      required:
      - reason
      properties:
        reason:
          type: string
...